Working with AlertControllers and making a more functional TableView
Know, I have a InitialTableViewController that has some section and,
datas inside. And I create alert controller for selection of any of the
table views cells. The alert has shows the table cells content as message
and has 2 buttons with title 'Ok' and 'Delete'. 'Ok' closes the alert and
delects the selected row cell. 'Delete' deletes the row from tableData and
uptades the tableView